added comments section

+++ client/src/components/Diary/Diary.jsx
@@ -9,10 +9,6 @@ function Diary(props){
 
   let history = useHistory();
 
-  const config = {
-    headers: { "Authorization": "Bearer " + props.uToken }
-  };
-
   const [title, setTitle] = useState("");
   const [content, setContent] = useState("");
   const [date, setDate] = useState("");
@@ -21,41 +17,74 @@ function Diary(props){
 
   useEffect(() => {
 
-    axios.get("http://localhost:5000/diary/alldays", config).then(res => {
-      setUserDays(() => {
-        return res.data.days.filter((day) => {
-          return day.user_id === props.uID;
+    const loggedInUser = localStorage.getItem("userData");
+
+    if (loggedInUser) {
+      const foundUser = JSON.parse(loggedInUser);
+
+      const config = {
+        headers: { "Authorization": "Bearer " + foundUser.token }
+      };
+
+      axios.get("http://localhost:5000/diary/alldays", config).then(res => {
+        setUserDays(() => {
+          return res.data.days.filter((day) => {
+            return day.user_id === props.uID;
+          });
         });
+      }).catch((error) => {
+        console.log(error.response.status);
+        history.push("/login");
       });
-    }).catch((error) => {
-      console.log(error.response.status);
+    }
+    else{
       history.push("/login");
-    });
+    }
   });
 
-  function deleteDay(id) {
-    axios.delete("http://localhost:5000/diary/day/delete/"  + id, config).then(res => {
-      console.log(res);
-    });
+  function deleteDay(id){
+    const loggedInUser = localStorage.getItem("userData");
 
-  setUserDays((preValues) => {
-    return userDays.filter((day, index) => {
-      return index !== id;
+    if (loggedInUser) {
+      const foundUser = JSON.parse(loggedInUser);
+
+      const config = {
+        headers: { "Authorization": "Bearer " + foundUser.token }
+      };
+
+      axios.delete("http://localhost:5000/diary/day/delete/"  + id, config).then(res => {
+        console.log(res);
+      });
+
+    setUserDays((preValues) => {
+      return userDays.filter((day, index) => {
+        return index !== id;
+      });
     });
-  });
+  }
 }
 
 function getData(id){
-  axios.get("http://localhost:5000/diary/" + id, config).then(res => {
-        setTitle(res.data.title);
-        setContent(res.data.content);
-        setDate(res.data.date);
-    });
+
+  const loggedInUser = localStorage.getItem("userData");
+  if (loggedInUser) {
+    const foundUser = JSON.parse(loggedInUser);
+
+    const config = {
+      headers: { "Authorization": "Bearer " + foundUser.token }
+    };
+
+    axios.get("http://localhost:5000/diary/" + id, config).then(res => {
+          setTitle(res.data.title);
+          setContent(res.data.content);
+          setDate(res.data.date);
+      });
+  }
 }
 
   return (
     <div>
-      <Header checkAuth={props.changeAuthStatus} hOption2="Home"/>
+      <Header hOption2="Home"/>
       <CreateArea userId={props.uID}
                   userName={props.uName}
                   uToken={props.uToken}
